requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
praw==7.7.1
supabase==2.3.4
google-generativeai==0.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
slack-sdk==3.26.1
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.4
pytest-asyncio==0.23.3
asyncpraw==7.7.1
